Prepping work on run command tests
