+++ src/run/opts.js
@@ -1,9 +1,14 @@
-module.exports = [
-  {
-    name: 'name',
-    type: 'list',
-    choices: funcs,
-    message: 'Which lambda function?',
-    when: () => !flags.name
-  }
-]
+const loadFuncs = require('../util/load-funcs')
+
+module.exports = function(){
+  const funcs = loadFuncs()
+
+  return [
+    {
+      name: 'name',
+      type: 'list',
+      choices: funcs,
+      message: 'Which lambda function?'
+    }
+  ]
+}
