+++ src/util/load-funcs.js
@@ -1,0 +1,6 @@
+const glob = require('glob')
+
+module.exports = function(){
+  return glob.sync('functions/*')
+  .map((path) => path.split('/').pop())
+}
